+++ MicroCore/Reflection/Types/MicroReflectField.cpp
@@ -49,7 +49,7 @@ namespace micro {
 
 	ReflectField::ReflectField(
 		micro_string name,
-		const size_t offset,
+		const uint32_t offset,
 		const ReflectType* type
 	)
 		: ReflectField{ name, offset, type, MicroReflectAccessor::Public }
@@ -57,7 +57,7 @@ namespace micro {
 
 	ReflectField::ReflectField(
 		micro_string name,
-		const size_t offset,
+		const uint32_t offset,
 		const ReflectType* type,
 		const MicroReflectAccessor accessor
 	)
